contracts update
